refactor(headless): split Dialog into per-file parts for tree-shaking
Split the monolithic dialog.tsx into individual files per subcomponent
and use `export * as Dialog` namespace re-export instead of Object.assign.
This enables bundlers to tree-shake unused Dialog parts at the module level.

API change: <Dialog> becomes <Dialog.Root>
